+++ js/form.js
@@ -1,6 +1,7 @@
 import {showSuccessMessage, showErrorMessage} from './messages.js';
 import {sendData} from './api.js';
 import {resetMapMainMarker} from './map.js';
+import {resetFiltersForm} from './filter.js';
 
 const form = document.querySelector('.ad-form');
 const selectedType = form.querySelector('#type');
@@ -96,6 +97,7 @@ timeoutField.addEventListener('change', ()=>{timeinField.value = timeoutField.va
 
 const resetForm = ()=>{
   form.reset();
+  resetFiltersForm();
   resetMapMainMarker();
 ///////////////////////////////
 };
