+++ js/map.js
@@ -1,7 +1,10 @@
-import {roundLatLng, showAlert} from './util.js';
+import {roundLatLng} from './util.js';
 import {getData} from './api.js';
 import {generateCard} from './generate-card.js';
 import {activatePage, deactivatePage} from './page-behavior.js';
+import {showAlert} from './messages.js';
+import {getFilteringAdverts} from './filter.js';
+
 const addressField = document.querySelector('#address');
 
 const MAIN_PIN_SIZE  = 52;
@@ -57,6 +60,7 @@ const resetMapMainMarker = ()=>{
 };
 
 const renderMarkersOnMap = (adverts)=>{
+  markersLayer.clearLayers();
   adverts.forEach((advert)=>{
     const popupCard = document.createElement('div');
     popupCard.append(generateCard(advert));
@@ -74,7 +78,7 @@ const renderMarkersOnMap = (adverts)=>{
 
 const onSuccessRequest = (adverts)=>{
   activatePage();
-  renderMarkersOnMap(adverts);
+  renderMarkersOnMap(getFilteringAdverts(adverts));
 };
 
 const onFailRequest = (err)=>{
